Update README.md
Updated to focus on the capacity of the dining halls rather than every space on campus. The first paragraph has been transformed completely while the second paragraph has a few changes to be more specific to the dining halls. The first paragraph feels like it could use another sentence at the end but not sure what to put yet..